% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_lot.R
\name{assign_lot}
\alias{assign_lot}
\title{Assign lines of therapy}
\usage{
assign_lot(
  .data,
  treatment = treatment,
  dt_trt_observed = dt_trt_observed,
  dt_treatment_end = dt_treatment_end
)
}
\arguments{
\item{.data}{Data frame with treatment information}

\item{treatment}{Column containing treatment names}

\item{dt_trt_observed}{Column containing treatment start date}

\item{dt_treatment_end}{Column containing treatment end date}
}
\value{
Data frame with assigned line of therapy and regimen information
}
\description{
This algorithm assigns treatments to the same line when they occur
within 30 days of an initial treatment in a line. Treatments
occurring greater than 30 days from line initiation begin a new line.
}
