% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_medidata}
\alias{dictionary_medidata}
\title{Medidata dictionary}
\format{
A data frame with 8 variables:
\describe{
\item{Table}{Medidata table}
\item{Table description}{Medidata table description}
\item{Variable}{Variable name in medidata table}
\item{Variable label}{If available, a descriptive label for the variable}
\item{R column type}{How the variable is stored in R}
\item{Prompt text}{Prompt text in the electronic data capture}
}
}
\usage{
dictionary_medidata
}
\description{
Data dictionary of clinical data recorded in medidata. Subject to change.
The data consists of multiple tables.
}
\examples{
head(dictionary_medidata)
}
\keyword{datasets}
