% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_proms}
\alias{dictionary_proms}
\title{PROMS dictionary}
\format{
A data frame with 8 variables:
\describe{
\item{instrument}{Survey instrument}
\item{instrument_description}{Survey instrument description}
\item{variable_type}{Type of survey question (one of multiple choice, select all that apply, write-in, etc.)}
\item{variable}{Variable name that references instrument and question}
\item{label}{Variable label with question text}
\item{values_original}{Reponse values as original programmed}
\item{values_clean}{Response values cleaned with lower case and extra spaces removed}
\item{col_type}{How the variable is stored in R}
}
}
\usage{
dictionary_proms
}
\description{
Data dictionary of Patient Reported Outcomes and Measures (PROMS) self-report survey data. Subject to change.
}
\examples{
head(dictionary_proms)
}
\keyword{datasets}
