% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_blind_medidata}
\alias{lst_blind_medidata}
\title{Blinded Sample Medidata}
\format{
A list of 7 data frames:
\describe{
\item{ic}{Consent table}
\item{ca_cm_derived}{Treatment table with mapping}
\item{cyc_date}{Visit Date Table}
\item{pq}{PQ version 1 & 2}
\item{new_pq}{PQ version 3}
\item{new_pq_v4}{PQ version 4}
\item{new_pq_v5}{PQ version 5}
}
}
\usage{
lst_blind_medidata
}
\description{
Randomized blinded sample of 100 subjects from medidata
}
\examples{
head(lst_blind_medidata)
}
\keyword{datasets}
