% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proms_q_codes}
\alias{proms_q_codes}
\title{PROMS question codes}
\format{
A data frame with 462 rows and 2 variables:
\describe{
\item{question_code}{Variable name of PROM question}
\item{question_text}{Full text of PROM question}
}
}
\source{
"TrueNTH Questionnaire Code Dictionary - IRONMAN_v2.0.xlsx"
}
\usage{
proms_q_codes
}
\description{
A dataset containing the variable names and corresponding question text for all PROMS.
The variables are as follows:
}
\keyword{datasets}
