% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proms_q_opts}
\alias{proms_q_opts}
\title{PROMS question options}
\format{
A data frame with 1099 rows and 4 variables:
\describe{
\item{Option Code}{ID for PROM answer}
\item{option_code}{Variable name of PROM question}
\item{option_order}{Ordering of PROM answers within a question}
\item{Option Text}{PROM text response}
}
}
\source{
"TrueNTH Questionnaire Code Dictionary - IRONMAN_v2.0.xlsx"
}
\usage{
proms_q_opts
}
\description{
A dataset containing the levels and possible answers of all PROM questions.
The variables are as follows:
}
\keyword{datasets}
