% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medidata.R
\name{read_medidata}
\alias{read_medidata}
\title{Read all .sas7bdat IRONMAN files from a directory}
\usage{
read_medidata(path, clean = TRUE, skip = "_raw")
}
\arguments{
\item{path}{A directory path typically created with fs::path()
or here::here() which contains the .sas7dbat files to read.}

\item{clean}{If TRUE (default), use janitor::clean_names() to
clean column names.}

\item{skip}{A character string, when detected in the file name,
that will cause matching files not to be read.}
}
\value{
A named list of tibbles, where the names correspond to
the name of the sas7bdat source file for each list element.
}
\description{
Data exports (e.g. from Medidata) are most often provided
as a directory with many .sas7bdat files. This is a
convenience function to read all such files in a given
directory and return as a named list of tibbles.
}
\examples{
\dontrun{
medidata <- read_medidata(
  fs::path(
    "~/Documents/gh-repos-mskcc/c16-170-ironman/data",
    "20210601/c16_170_IRONMAN_731_20210601_171659"
  ),
  skip = "y|m|r|o|g|p|c" #reads only 9 of the files
)

# view 3 columns from each tibble
medidata \%>\%
  purrr::map(~ .x \%>\% dplyr::select(project, site, instance_name))
  }
}
