% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_to_country}
\alias{site_to_country}
\title{Site to Country mapping}
\format{
A data frame with 180 rows and 3 variables:
\describe{
\item{site_id}{Numeric site id}
\item{site_code}{Character site code with two digit padded zero}
\item{site_name}{Full site name}
\item{site_name_short}{Short site name}
\item{site_name_true_nth}{Site name recorded in True North / PROMS}
\item{site_country}{Country of site}
}
}
\source{
"PCCTC internal documentation"
}
\usage{
site_to_country
}
\description{
A data set of all PCCTC site IDs mapping from treatment site to country.
Not all IDs apply to Ironman.
Note that as sites get added this data may need updating; if so, please
file an issue or contact PCCTC. In addition,
curated data deliveries now contain country information
in the medidata subject table.
}
\details{
In addition, when identifying the site of a subject, it is best practice
to use actual site fields, rather than extracting from subject id. The site
identifier used in the subject id represents the site at the time of enrollment,
and subjects may change sites over time. The curent site is reflected in the
site fields, where as the enrollment site is reflected in the subject id.
}
\keyword{datasets}
