% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_map}
\alias{treatment_map}
\title{Treatment mapping}
\format{
A data frame with 5 variables:
\describe{
\item{treatment}{Treatment as submitted to the study}
\item{treatment_lower}{Treatment as submitted to the study lower case}
\item{treatment_recode}{Assigned treatment standardized coding}
\item{treatment_category}{Assigned treatment category}
\item{is_taxel}{Indicates if treatment is a \code{taxel} (one of Cabazitaxel, Docetaxel, Paclitaxel, Tesetaxel)}
\item{is_nha}{Indicates if treatment is a novel hormone agent (one of Abiraterone, Enzalutamide, Darolutamide, Apalutamide)}
}
}
\usage{
treatment_map
}
\description{
A dataset containing treatment mapping specific to PCCTC studies, intended
to be updated regularly. If a mapping needs to be corrected or added,
please file an issue or contact PCCTC. Note that this mapping should already
be in the \code{priorrx_derived} and \code{treatment_derived} tables.
}
\examples{
print(treatment_map)
}
\keyword{datasets}
