% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_medidata}
\alias{dictionary_medidata}
\title{Medidata dictionary}
\format{
A data frame with 8 variables:
\describe{
\item{table}{Medidata table}
\item{table_description}{Medidata table description}
\item{variable}{Variable name in medidata table}
\item{variable_label}{If available, a descriptive label for the variable}
\item{r_column_type}{How the variable is stored in R}
\item{drs_prompt_text}{Design Request Specification: Prompt text in the electronic data capture}
\item{drs_format}{Design Request Specification: Variable capture SAS format specification electronic data capture}
\item{drs_entry_type}{Design Request Specification: Variable field type in electronic data capture}
}
}
\usage{
dictionary_medidata
}
\description{
Data dictionary of clinical data recorded in medidata. Subject to change.
The data consists of multiple tables.
}
\examples{
head(dictionary_medidata)
}
\keyword{datasets}
